requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements-dev.txt
#
#-e file:///C:/ProgramData/workspace/panopto-watershed-cron
    # via -r requirements-dev.txt
appdirs==1.4.3
    # via zeep
attrs==17.3.0
    # via
    #   pytest
    #   zeep
bcrypt==3.2.0
    # via paramiko
cached-property==1.3.1
    # via zeep
certifi==2017.11.5
    # via requests
cffi==1.15.0
    # via
    #   bcrypt
    #   cryptography
    #   pynacl
charset-normalizer==2.0.12
    # via requests
click==8.0.0
    # via pip-tools
colorama==0.3.9
    # via
    #   click
    #   pytest
coverage==4.4.2
    # via
    #   pip-conflict-checker
    #   pytest-cov
cryptography==36.0.1
    # via paramiko
defusedxml==0.5.0
    # via zeep
first==2.0.1
    # via pip-tools
flake8==3.5.0
    # via
    #   -r requirements-dev.txt
    #   pip-conflict-checker
idna==2.6
    # via requests
isodate==0.6.0
    # via zeep
lxml==4.8.0
    # via zeep
mccabe==0.6.1
    # via flake8
mock==2.0.0
    # via pip-conflict-checker
nose==1.3.7
    # via pip-conflict-checker
paramiko==2.9.2
    # via panopto-python-soap
pbr==3.1.1
    # via mock
pip-conflict-checker==0.5.0
    # via -r requirements-dev.txt
pip-tools==1.10.2
    # via -r requirements-dev.txt
pluggy==0.6.0
    # via pytest
py==1.5.2
    # via pytest
pycodestyle==2.3.1
    # via flake8
pycparser==2.21
    # via cffi
pycron==3.0.0
    # via panopto-python-soap
pyflakes==1.6.0
    # via flake8
pynacl==1.5.0
    # via paramiko
pytest==3.3.0
    # via
    #   -r requirements-dev.txt
    #   pytest-cov
pytest-cov==2.5.1
    # via -r requirements-dev.txt
pytest-runner==3.0
    # via -r requirements-dev.txt
python-dotenv==0.19.2
    # via panopto-python-soap
pytz==2017.3
    # via
    #   panopto-python-soap
    #   zeep
regex==2022.1.18
    # via panopto-python-soap
requests==2.27.1
    # via
    #   requests-toolbelt
    #   zeep
requests-toolbelt==0.8.0
    # via zeep
six==1.11.0
    # via
    #   bcrypt
    #   isodate
    #   mock
    #   pip-tools
    #   pytest
    #   zeep
urllib3==1.24.2
    # via
    #   panopto-python-soap
    #   requests
zeep==3.1.0
    # via panopto-python-soap

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
